+++ pkg/api/api.go
@@ -236,6 +236,13 @@ const (
 	LogLevelError
 )
 
+type LogStyle uint8
+
+const (
+	LogStyleDefault LogStyle = iota
+	LogStyleVisualStudio
+)
+
 type Charset uint8
 
 const (
@@ -280,6 +287,7 @@ const (
 type BuildOptions struct {
 	Color       StderrColor         // Documentation: https://esbuild.github.io/api/#color
 	LogLevel    LogLevel            // Documentation: https://esbuild.github.io/api/#log-level
+	LogStyle    LogStyle            // Documentation: https://esbuild.github.io/api/#log-style
 	LogLimit    int                 // Documentation: https://esbuild.github.io/api/#log-limit
 	LogOverride map[string]LogLevel // Documentation: https://esbuild.github.io/api/#log-override
 	AbsPaths    AbsPaths            // Documentation: https://esbuild.github.io/api/#abs-path
@@ -412,6 +420,7 @@ func Build(options BuildOptions) BuildResult {
 type TransformOptions struct {
 	Color       StderrColor         // Documentation: https://esbuild.github.io/api/#color
 	LogLevel    LogLevel            // Documentation: https://esbuild.github.io/api/#log-level
+	LogStyle    LogStyle            // Documentation: https://esbuild.github.io/api/#log-style
 	LogLimit    int                 // Documentation: https://esbuild.github.io/api/#log-limit
 	LogOverride map[string]LogLevel // Documentation: https://esbuild.github.io/api/#log-override
 	AbsPaths    AbsPaths            // Documentation: https://esbuild.github.io/api/#abs-path
@@ -717,6 +726,7 @@ type FormatMessagesOptions struct {
 	TerminalWidth int
 	Kind          MessageKind
 	Color         bool
+	LogStyle      LogStyle
 }
 
 func FormatMessages(msgs []Message, opts FormatMessagesOptions) []string {